+++ django_fixmystreet/fixmystreet/forms.py
@@ -133,7 +133,6 @@ class CitizenForm(forms.Form):
             instance = FMSUser.objects.get(email=self.cleaned_data["email"]);
         except FMSUser.DoesNotExist:
             data = self.cleaned_data.copy()
-            del data['subscription']
             #For unique constraints
             data['username'] = data['email']
             instance = FMSUser.objects.create(**data)