+++ django_fixmystreet/fixmystreet/static/js/new_report.js
@@ -60,7 +60,7 @@ function retrieveAddress() {
         $('#address-text').removeClass('loading');
         $form.find('.text-error').remove();
 
-        if (!(address.street.postCode in zipcodes) || !zipcodes[String(address.street.postCode)].participation) {
+        if (typeof window.zipcodes != 'undefined' && address.street.postCode in zipcodes && !zipcodes[String(address.street.postCode)].participation) {
 
             fillAdressField(lang, address);
             //This commune does not participate to fixmystreet until now.