+++ django_fixmystreet/fixmystreet/tests/__init__.py
@@ -1,8 +1,8 @@
-import shutil
-import os
+
 
 from django.core.files.storage import default_storage
 from django.test import TestCase
+
 
 class SampleFilesTestCase(TestCase):
     fixtures = ['bootstrap', 'sample']
@@ -10,25 +10,32 @@ class SampleFilesTestCase(TestCase):
     @classmethod
     def setUpClass(cls):
         default_storage.location = 'media/' # force using source media folder to avoid real data erasing
-    # @classmethod
-    # def setUpClass(cls):
-        # shutil.copytree('media', 'media-tmp')
-        # default_storage.location = 'media-tmp'
-# 
-    # @classmethod
-    # def tearDownClass(self):
-        # shutil.rmtree('media-tmp')
-
-    # def _fixture_setup(self):
-    #     if os.path.exists('media/photos'):
-    #         shutil.rmtree('media/photos')
-    #     shutil.copytree('media/photos-sample', 'media/photos')
-    #     super(SampleFilesTestCase, self)._fixture_setup()
-
 
 from django_fixmystreet.fixmystreet.tests.views import *
 from django_fixmystreet.fixmystreet.tests.reports import *
 from django_fixmystreet.fixmystreet.tests.users import *
 from django_fixmystreet.fixmystreet.tests.organisation_entity import *
 from django_fixmystreet.fixmystreet.tests.mail import *
-#from django_fixmystreet.fixmystreet.tests.api import *
+
+# class TestDiscoveryPlugin(Plugin):
+#     enabled = True
+
+#     def configure(self, options, conf):
+#         super(TestDiscoveryPlugin, self).configure(options, conf)
+#         self.enabled = True
+
+#     def wantDirectory(self, dirname):
+#         parts = dirname.split(os.path.sep)
+#         print parts
+#         return 'tests' in parts
+
+#     def wantClass(self, cls):
+#         return issubclass(cls, TestCase)
+
+#     def wantFile(self, filename):
+#         parts = filename.split(os.path.sep)
+#         return 'tests' in parts and filename.endswith('.py')
+
+#     def wantModule(self, module):
+#         parts = module.__name__.split('.')
+#         return 'tests' in parts