+++ django_fixmystreet/fixmystreet/tests/mail.py
@@ -43,6 +43,7 @@ class MailTest(TestCase):
 			'report-postalcode':'1210',
 			'report-category':'1',
 			'report-secondary_category':'1',
+			'report-subscription':'on',
 			'citizen-quality':'1',
 			'comment-text':'test',
 			'files-TOTAL_FORMS': 0,
@@ -51,7 +52,6 @@ class MailTest(TestCase):
 			'citizen-email':self.citizen.email,
 			'citizen-firstname':self.citizen.first_name,
 			'citizen-lastname':self.citizen.last_name,
-			'citizen-subscription':'on',
 		}
 
 	def testCreateReportMail(self):
@@ -73,27 +73,34 @@ class MailTest(TestCase):
 		self.assertEquals(response.status_code, 200)
 		# self.assertIn('/en/report/trou-en-revetements-en-trottoir-en-saint-josse-ten-noode/1', response['Location'])
 
+
 		# report_id = resolve(response.redirect_chain[-1][0]).kwargs['report_id']
-		report_id = 1
+		report = Report.objects.get(id=1)
+		self.assertEquals(report.subscriptions.all().count(), 2)
+		self.assertEquals(len(mail.outbox), 2)
+
 		#Login to access the pro page to create a user
 		self.client.login(username='manager@a.com', password='test')
 
 		#Accept the created report
-		response = self.client.get(reverse('report_accept_pro', args=[report_id]), follow=True)
+		response = self.client.get(reverse('report_accept_pro', args=[report.id]), follow=True)
 		#The status of the report must now be MANAGER_ASSIGNED
 		self.assertEquals(response.status_code, 200)
-		self.assertTrue(Report.objects.get(pk=1).status == Report.MANAGER_ASSIGNED)
-		#3 mails have been sent, 2 for the report creation, 1 for the report publishing
-		self.assertEquals(len(mail.outbox),3)
+
+		report = Report.objects.get(id=report.id)
+		self.assertEquals(report.status, Report.MANAGER_ASSIGNED)
+		#3 mails have been sent, 2 for the report creation + 1 notification to author for the report publishing
+		self.assertEquals(len(mail.outbox), 3)
 		#Close the report
-		response = self.client.get(reverse('report_close_pro', args=[report_id]), follow=True)
+		response = self.client.get(reverse('report_close_pro', args=[report.id]), follow=True)
 		self.assertEquals(response.status_code, 200)
-		#The status of the report must now be PROCESSED
-		self.assertTrue(Report.objects.get(pk=1).status == Report.PROCESSED)
+
+		report = Report.objects.get(id=report.id)
+		self.assertEquals(report.status, Report.PROCESSED)
 		#4 mails have been sent, 2 for the report creation, 1 for the report publishing and 1 for closing the report
-		self.assertEquals(len(mail.outbox),4)
+		self.assertEquals(len(mail.outbox), 5)
 		#The last one must be sent to the citizen (= the closing report mail)
-		self.assertTrue(self.citizen.email in mail.outbox[2].to)
+		self.assertIn(self.citizen.email, mail.outbox[4].to)
 
 
 	def testRefuseReportMail(self):
@@ -102,7 +109,10 @@ class MailTest(TestCase):
 		self.assertEquals(response.status_code, 200)
 		# self.assertIn('/en/report/trou-en-revetements-en-trottoir-en-saint-josse-ten-noode/1', response['Location'])
 
-		self.assertEquals(len(mail.outbox),2) # one for creator subscription, one for manager
+		report = Report.objects.get(id=1)
+		self.assertEquals(report.subscriptions.all().count(), 2)
+		self.assertEquals(len(mail.outbox), 2) # one for creator subscription, one for manager
+
 		#Login to access the pro page to create a user
 		self.client.login(username='manager@a.com', password='test')
 		#Refuse the created report
@@ -110,10 +120,11 @@ class MailTest(TestCase):
 		self.assertEquals(response.status_code, 200)
 		self.assertEquals(len(mail.outbox),3)
 		#The status of the report must now be REFUSED
-		self.assertTrue(Report.objects.get(pk=1).status == Report.REFUSED)
+		report = Report.objects.get(id=report.id)
+		self.assertEquals(report.status, Report.REFUSED)
 		#3 mails have been sent, 2 for the report creation and 1 for refusing the report
 		#The last one must be sent to the citizen (= the refusing report mail)
-		self.assertTrue(self.citizen.email in mail.outbox[2].to)
+		self.assertIn(self.citizen.email, mail.outbox[2].to)
 
 	def testSubscriptionForCititzenMail(self):
 		#Send a post request filling in the form to create a report
@@ -122,6 +133,7 @@ class MailTest(TestCase):
 		# self.assertIn('/en/report/trou-en-revetements-en-trottoir-sen-saint-josse-ten-noode/1', response['Location'])
 
 		self.assertEquals(len(mail.outbox),2) # one for creator subscription, one for manager
+
 		#Send a post request subscribing a citizen to the just created report
 		response = self.client.post('/en/report/1/subscribe/',{'citizen_email':'post@test.com'}, follow=True)
 		self.assertEquals(response.status_code, 200)
@@ -135,13 +147,14 @@ class MailTest(TestCase):
 		self.assertEquals(response.status_code, 200)
 		# self.assertIn('/en/report/trou-en-revetements-en-trottoir-en-saint-josse-ten-noode/1', response['Location'])
 
-		self.assertEquals(len(mail.outbox),2) # one for creator subscription, one for manager
+		self.assertEquals(len(mail.outbox), 2) # one for creator subscription, one for manager
 
 		#Login to access the pro page
 		self.client.login(username='manager@a.com', password='test')
 		#Publish the created report
 		response = self.client.post('/en/pro/report/1/accept/', follow=True)
 		self.assertEquals(response.status_code, 200)
+		self.assertEquals(len(mail.outbox), 3)
 
 		#Send a post request to mark the report as done
 
@@ -151,12 +164,12 @@ class MailTest(TestCase):
 		#4 mails have been sent, 2 for the report creation and 1 for telling the responsible manager that the report is marked as done, and 1 for the report change to the citizen subscriber
 		self.assertEquals(Report.objects.get(id=1).status, Report.SOLVED)
 
-		self.assertEquals(len(mail.outbox),4)
+		self.assertEquals(len(mail.outbox), 4)
 		self.assertTrue(self.manager.email in mail.outbox[3].to)
 		#Send another post request to mark the report as done
-		response = self.client.post('/en/report/1/update/',{'is_fixed':'True'})
+		response = self.client.post('/en/report/1/update/', {'is_fixed':'True'})
 		self.assertEquals(response.status_code, 302)
 		# self.assertIn('/en/report/trou-en-revetements-en-trottoir-en-saint-josse-ten-noode/1', response['Location'])
 		#Again 4 mails have been sent, the extra mark as done request will not send an extra email to the responsible manager
-		self.assertEquals(len(mail.outbox),4)
+		self.assertEquals(len(mail.outbox), 4)